Today I have completed the initial stages of preparing a map product and the contained data.
My code so far has generated matplotlib handles for setting up a legend, creation of a 25Km scale bar with 5Km labelled increments. The datasets for extent and display features were loaded and the correct coordinate system set.

I have encountered a confusing issued that needs further investigation related to the ax.set_extent. With UTM it should be in metres, but I can only get the plot to display at an appropriate zoom level with values of 0.1. This should be in metres so likely a coordinate system issue?

Lastly a background colour was defined